contents added for twitter card #1107
