+++ lib/app/shared/constants/urls.dart
@@ -27,6 +27,8 @@ class Urls {
   static const String chainbornMembershipCardUrl =
       'https://issuer.talao.co/chainborn/membershipcard/';
 
+  static const String twitterCardUrl = 'https://issuer.talao.co/twitter/';
+
   static const String identityCardUrl = 'http://issuer.talao.co/kyc';
   static const String talaoIpfsGateway = 'https://talao.mypinata.cloud/ipfs/';
