+++ lib/app/shared/dio_client/dio_client.dart
@@ -81,7 +81,8 @@ class DioClient {
     Map<String, dynamic> headers,
   ) async {
     if (uri.contains(Urls.tezotopiaMembershipCardUrl) ||
-        uri.contains(Urls.chainbornMembershipCardUrl)) {
+        uri.contains(Urls.chainbornMembershipCardUrl) ||
+        uri.contains(Urls.twitterCardUrl)) {
       await dotenv.load();
       final YOTI_AI_API_KEY = dotenv.get('YOTI_AI_API_KEY');
