+++ lib/app/shared/enum/message/response_string/response_string.dart
@@ -79,6 +79,9 @@ enum ResponseString {
   RESPONSE_STRING_chainbornMembershipWhyGetThisCard,
   RESPONSE_STRING_chainbornMembershipExpirationDate,
   RESPONSE_STRING_chainbornMembershipHowToGetIt,
+  RESPONSE_STRING_twitterWhyGetThisCard,
+  RESPONSE_STRING_twitterExpirationDate,
+  RESPONSE_STRING_twitterHowToGetIt,
   RESPONSE_STRING_tezoniaPassWhyGetThisCard,
   RESPONSE_STRING_tezoniaPassExpirationDate,
   RESPONSE_STRING_tezoniaPassHowToGetIt,
