+++ lib/app/shared/enum/type/credential_subject_type/credential_subject_type.dart
@@ -51,224 +51,5 @@ enum CredentialSubjectType {
   aragoLearningAchievement,
   aragoOver18,
   pcdsAgentCertificate,
-}
-
-extension CredentialSubjectTypeX on CredentialSubjectType {
-  bool isDisabled() {
-    if (this == CredentialSubjectType.dogamiPass) {
-      return true;
-    } else if (this == CredentialSubjectType.pigsPass) {
-      return true;
-    } else if (this == CredentialSubjectType.bunnyPass) {
-      return true;
-    } else if (this == CredentialSubjectType.troopezPass) {
-      return true;
-    } else if (this == CredentialSubjectType.tzlandPass) {
-      return true;
-    } else if (this == CredentialSubjectType.matterlightPass) {
-      return true;
-    } else if (this == CredentialSubjectType.tezoniaPass) {
-      return true;
-    } else {
-      return false;
-    }
-  }
-
-  String get name {
-    switch (this) {
-      case CredentialSubjectType.bloometaPass:
-        return 'BloometaPass';
-      case CredentialSubjectType.tezoniaPass:
-        return 'TezoniaPass';
-      case CredentialSubjectType.tzlandPass:
-        return 'TzlandPass';
-      case CredentialSubjectType.troopezPass:
-        return 'TroopezPass';
-      case CredentialSubjectType.pigsPass:
-        return 'PigsPass';
-      case CredentialSubjectType.matterlightPass:
-        return 'MatterlightPass';
-      case CredentialSubjectType.dogamiPass:
-        return 'DogamiPass';
-      case CredentialSubjectType.bunnyPass:
-        return 'BunnyPass';
-      case CredentialSubjectType.tezotopiaMembership:
-        return 'MembershipCard_1';
-      case CredentialSubjectType.chainbornMembership:
-        return 'Chainborn_MembershipCard';
-      case CredentialSubjectType.ageRange:
-        return 'AgeRange';
-      case CredentialSubjectType.nationality:
-        return 'Nationality';
-      case CredentialSubjectType.gender:
-        return 'Gender';
-      case CredentialSubjectType.deviceInfo:
-        return 'DeviceInfo';
-      case CredentialSubjectType.tezosAssociatedWallet:
-        return 'TezosAssociatedAddress';
-      case CredentialSubjectType.ethereumAssociatedWallet:
-        return 'EthereumAssociatedAddress';
-      case CredentialSubjectType.fantomAssociatedWallet:
-        return 'FantomAssociatedAddress';
-      case CredentialSubjectType.polygonAssociatedWallet:
-        return 'PolygonAssociatedAddress';
-      case CredentialSubjectType.binanceAssociatedWallet:
-        return 'BinanceAssociatedAddress';
-      case CredentialSubjectType.certificateOfEmployment:
-        return 'CertificateOfEmployment';
-      case CredentialSubjectType.ecole42LearningAchievement:
-        return 'Ecole42LearningAchievement';
-      case CredentialSubjectType.emailPass:
-        return 'EmailPass';
-      case CredentialSubjectType.identityPass:
-        return 'IdentityPass';
-      case CredentialSubjectType.identityCard:
-        return 'IdCard';
-      case CredentialSubjectType.learningAchievement:
-        return 'LearningAchievement';
-      case CredentialSubjectType.loyaltyCard:
-        return 'LoyaltyCard';
-      case CredentialSubjectType.over18:
-        return 'Over18';
-      case CredentialSubjectType.over13:
-        return 'Over13';
-      case CredentialSubjectType.passportFootprint:
-        return 'PassportNumber';
-      case CredentialSubjectType.phonePass:
-        return 'PhoneProof';
-      case CredentialSubjectType.professionalExperienceAssessment:
-        return 'ProfessionalExperienceAssessment';
-      case CredentialSubjectType.professionalSkillAssessment:
-        return 'ProfessionalSkillAssessment';
-      case CredentialSubjectType.professionalStudentCard:
-        return 'ProfessionalStudentCard';
-      case CredentialSubjectType.residentCard:
-        return 'ResidentCard';
-      case CredentialSubjectType.selfIssued:
-        return 'SelfIssued';
-      case CredentialSubjectType.studentCard:
-        return 'StudentCard';
-      case CredentialSubjectType.voucher:
-        return 'Voucher';
-      case CredentialSubjectType.tezVoucher:
-        return 'TezVoucher_1';
-      case CredentialSubjectType.talaoCommunityCard:
-        return 'TalaoCommunity';
-      case CredentialSubjectType.diplomaCard:
-        return 'VerifiableDiploma';
-      case CredentialSubjectType.aragoPass:
-        return 'AragoPass';
-      case CredentialSubjectType.aragoEmailPass:
-        return 'AragoEmailPass';
-      case CredentialSubjectType.aragoIdentityCard:
-        return 'AragoIdCard';
-      case CredentialSubjectType.aragoLearningAchievement:
-        return 'AragoLearningAchievement';
-      case CredentialSubjectType.aragoOver18:
-        return 'AragoOver18';
-      case CredentialSubjectType.pcdsAgentCertificate:
-        return 'PCDSAgentCertificate';
-      case CredentialSubjectType.defaultCredential:
-        return '';
-    }
-  }
-
-  CredentialSubjectModel modelFromJson(Map<String, dynamic> json) {
-    switch (this) {
-      case CredentialSubjectType.bloometaPass:
-        return BloometaPassModel.fromJson(json);
-      case CredentialSubjectType.tezoniaPass:
-        return TezoniaPassModel.fromJson(json);
-      case CredentialSubjectType.tzlandPass:
-        return TzlandPassModel.fromJson(json);
-      case CredentialSubjectType.troopezPass:
-        return TroopezPassModel.fromJson(json);
-      case CredentialSubjectType.pigsPass:
-        return PigsPassModel.fromJson(json);
-      case CredentialSubjectType.matterlightPass:
-        return MatterlightPassModel.fromJson(json);
-      case CredentialSubjectType.dogamiPass:
-        return DogamiPassModel.fromJson(json);
-      case CredentialSubjectType.bunnyPass:
-        return BunnyPassModel.fromJson(json);
-      case CredentialSubjectType.tezotopiaMembership:
-        return TezotopiaMembershipModel.fromJson(json);
-      case CredentialSubjectType.chainbornMembership:
-        return ChainbornMembershipModel.fromJson(json);
-      case CredentialSubjectType.ageRange:
-        return AgeRangeModel.fromJson(json);
-      case CredentialSubjectType.nationality:
-        return NationalityModel.fromJson(json);
-      case CredentialSubjectType.gender:
-        return GenderModel.fromJson(json);
-      case CredentialSubjectType.deviceInfo:
-        return DeviceInfoModel.fromJson(json);
-      case CredentialSubjectType.tezosAssociatedWallet:
-        return TezosAssociatedAddressModel.fromJson(json);
-      case CredentialSubjectType.ethereumAssociatedWallet:
-        return EthereumAssociatedAddressModel.fromJson(json);
-      case CredentialSubjectType.fantomAssociatedWallet:
-        return FantomAssociatedAddressModel.fromJson(json);
-      case CredentialSubjectType.polygonAssociatedWallet:
-        return PolygonAssociatedAddressModel.fromJson(json);
-      case CredentialSubjectType.binanceAssociatedWallet:
-        return BinanceAssociatedAddressModel.fromJson(json);
-      case CredentialSubjectType.certificateOfEmployment:
-        return CertificateOfEmploymentModel.fromJson(json);
-      case CredentialSubjectType.defaultCredential:
-        return DefaultCredentialSubjectModel.fromJson(json);
-      case CredentialSubjectType.ecole42LearningAchievement:
-        return Ecole42LearningAchievementModel.fromJson(json);
-      case CredentialSubjectType.emailPass:
-        return EmailPassModel.fromJson(json);
-      case CredentialSubjectType.identityPass:
-        return IdentityPassModel.fromJson(json);
-      case CredentialSubjectType.identityCard:
-        return IdentityCardModel.fromJson(json);
-      case CredentialSubjectType.learningAchievement:
-        return LearningAchievementModel.fromJson(json);
-      case CredentialSubjectType.loyaltyCard:
-        return LoyaltyCardModel.fromJson(json);
-      case CredentialSubjectType.over18:
-        return Over18Model.fromJson(json);
-      case CredentialSubjectType.over13:
-        return Over13Model.fromJson(json);
-      case CredentialSubjectType.passportFootprint:
-        return PassportFootprintModel.fromJson(json);
-      case CredentialSubjectType.phonePass:
-        return PhonePassModel.fromJson(json);
-      case CredentialSubjectType.professionalExperienceAssessment:
-        return ProfessionalExperienceAssessmentModel.fromJson(json);
-      case CredentialSubjectType.professionalSkillAssessment:
-        return ProfessionalSkillAssessmentModel.fromJson(json);
-      case CredentialSubjectType.professionalStudentCard:
-        return ProfessionalStudentCardModel.fromJson(json);
-      case CredentialSubjectType.residentCard:
-        return ResidentCardModel.fromJson(json);
-      case CredentialSubjectType.selfIssued:
-        return SelfIssuedModel.fromJson(json);
-      case CredentialSubjectType.studentCard:
-        return StudentCardModel.fromJson(json);
-      case CredentialSubjectType.voucher:
-        return VoucherModel.fromJson(json);
-      case CredentialSubjectType.tezVoucher:
-        return TezotopiaVoucherModel.fromJson(json);
-      case CredentialSubjectType.talaoCommunityCard:
-        return TalaoCommunityCardModel.fromJson(json);
-      case CredentialSubjectType.diplomaCard:
-        return DiplomaCardModel.fromJson(json);
-      case CredentialSubjectType.aragoPass:
-        return AragoPassModel.fromJson(json);
-      case CredentialSubjectType.aragoEmailPass:
-        return AragoEmailPassModel.fromJson(json);
-      case CredentialSubjectType.aragoIdentityCard:
-        return AragoIdentityCardModel.fromJson(json);
-      case CredentialSubjectType.aragoLearningAchievement:
-        return AragoLearningAchievementModel.fromJson(json);
-      case CredentialSubjectType.aragoOver18:
-        return AragoOver18Model.fromJson(json);
-      case CredentialSubjectType.pcdsAgentCertificate:
-        return PcdsAgentCertificateModel.fromJson(json);
-    }
-  }
+  twitterCard,
 }
