+++ lib/dashboard/home/tab_bar/credentials/models/model.dart
@@ -59,5 +59,6 @@ export 'tezos_associated_address/tezos_associated_address_model.dart';
 export 'tezotopia_membership/tezotopia_membership_model.dart';
 export 'tezotopia_voucher/tezotopia_voucher_model.dart';
 export 'troopez_pass/troopez_pass_model.dart';
+export 'twitter_card/twitter_card_model.dart';
 export 'tzland_pass/tzland_pass_model.dart';
 export 'voucher/voucher_model.dart';
