+++ lib/dashboard/home/tab_bar/credentials/widgets/display_in_list.dart
@@ -203,6 +203,8 @@ class DisplayInList extends StatelessWidget {
         return BinanceAssociatedAddressDisplayInList(
           credentialModel: credentialModel,
         );
+      case CredentialSubjectType.twitterCard:
+        return TwitterCardDisplayInList(credentialModel: credentialModel);
     }
   }
 }
