+++ lib/l10n/arb/app_en.arb
@@ -571,9 +571,12 @@
   "emailPassHowToGetIt": "It’s super easy. Altme will verify your email ownership by sending a code by email.",
   "tezotopiaMembershipWhyGetThisCard": "This Membership card will give you 25% cash backs on ALL Tezotopia Game transactions when you buy a Drops on the marketplace or mint an NFT on starbase.",
   "tezotopiaMembershipExpirationDate": "This card will remain active and reusable for 1 YEAR.",
-  "chainbornMembershipHowToGetIt": " To get this card, you need to share a proof of email. If you do not have it already, you can get this proof of email in Altme 'Discover' section.",
+  "chainbornMembershipHowToGetIt": "To get this card, you need to share a proof of email. If you do not have it already, you can get this proof of email in Altme 'Discover' section.",
   "chainbornMembershipWhyGetThisCard": "Be among the few that have access to exclusive Chainborn store content, airdrops and other member-only benefits !",
   "chainbornMembershipExpirationDate": "This card will remain active and reusable for 1 YEAR.",
+  "twitterHowToGetIt": "",
+  "twitterWhyGetThisCard": "",
+  "twitterExpirationDate": "",
   "tezotopiaMembershipHowToGetIt": "You need to present a Nationality Proof and an Age Range Proof. Claim them by following Altme’s KYC check.",
   "over18WhyGetThisCard": "This proof may be required by some Web 3 Apps / Websites to access their service or claim benefits : Membership card, Loyalty card, Rewards, etc.",
   "over18ExpirationDate": "This card will remain active and reusable for 1 YEAR.",
