+++ lib/l10n/untranslated.json
@@ -517,6 +517,9 @@
     "chainbornMembershipHowToGetIt",
     "chainbornMembershipWhyGetThisCard",
     "chainbornMembershipExpirationDate",
+    "twitterHowToGetIt",
+    "twitterWhyGetThisCard",
+    "twitterExpirationDate",
     "tezotopiaMembershipHowToGetIt",
     "over18WhyGetThisCard",
     "over18ExpirationDate",
@@ -1217,6 +1220,9 @@
     "chainbornMembershipHowToGetIt",
     "chainbornMembershipWhyGetThisCard",
     "chainbornMembershipExpirationDate",
+    "twitterHowToGetIt",
+    "twitterWhyGetThisCard",
+    "twitterExpirationDate",
     "tezotopiaMembershipHowToGetIt",
     "over18WhyGetThisCard",
     "over18ExpirationDate",
@@ -1917,6 +1923,9 @@
     "chainbornMembershipHowToGetIt",
     "chainbornMembershipWhyGetThisCard",
     "chainbornMembershipExpirationDate",
+    "twitterHowToGetIt",
+    "twitterWhyGetThisCard",
+    "twitterExpirationDate",
     "tezotopiaMembershipHowToGetIt",
     "over18WhyGetThisCard",
     "over18ExpirationDate",
@@ -2617,6 +2626,9 @@
     "chainbornMembershipHowToGetIt",
     "chainbornMembershipWhyGetThisCard",
     "chainbornMembershipExpirationDate",
+    "twitterHowToGetIt",
+    "twitterWhyGetThisCard",
+    "twitterExpirationDate",
     "tezotopiaMembershipHowToGetIt",
     "over18WhyGetThisCard",
     "over18ExpirationDate",
